Main changes included custom post types for the sidebar pages
Wordpress allows custom sidebar and those sidebars can be created vis
the register_sidebar function. more info can be found at:
http://codex.wordpress.org/Customizing_Your_Sidebar  In index.php I am
now calling the 2 different sidebars by their custom types.  I also
added custom post types for the latest sermon sections and stories
section, they will show up in the options pane within the dashboard. The
video type for the latest sermon works. All that is needed to embed it
is to go to Vimeo and grab the embed code and put that in the post's
content section.